successfully created admin controller
the admin controller works fine with the admin route. can be merged after being reviewed.
