+++ controller/admin.controller.js
@@ -1,4 +1,17 @@
+const express = require('express');
+const { restrict, checkAccess } = require('../lib/auth');
 const escape = require('html-entities').AllHtmlEntities;
+const colors = require('colors');
+const moment = require('moment');
+const fs = require('fs');
+const path = require('path');
+const multer = require('multer');
+const mime = require('mime-type/with-db');
+const csrf = require('csurf');
+const util = require('util');
+const stream = require('stream');
+const { validateJson } = require('../lib/schema');
+
 const{clearSessionValue,
      mongoSanitize,
      getThemes, 
@@ -20,8 +33,6 @@ const {
     deleteMenu,
     orderMenu
 } = require('../lib/menu');
-
-
 
 // Admin section
 const adminDashboard=( (req, res, next)=>{
@@ -113,7 +124,7 @@ const adminSetup=(async (req, res) => {
 
 
 // insert a user 5
-const setupUser=(async (req, res) => {   //can convert this to function because it seems the flow of control not entring try catch
+const setupUser=(async (req, res) => {   
     const db = req.app.db;
 
     const doc = {
@@ -261,6 +272,610 @@ const settingsMenu=(async (req, res) => {
     });
 });
 
+// page list
+const getPages=(async (req, res) => {
+    const db = req.app.db;
+    const pages = await db.pages.find({}).toArray();
+
+    res.render('settings-pages', {
+        title: 'Static pages',
+        pages: pages,
+        session: req.session,
+        admin: true,
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        config: req.app.config,
+        menu: sortMenu(await getMenu(db)),
+        csrfToken: req.csrfToken()
+    });
+});
+
+// pages new
+const newPages=(async (req, res) => {
+    const db = req.app.db;
+
+    res.render('settings-page', {
+        title: 'Static pages',
+        session: req.session,
+        admin: true,
+        button_text: 'Create',
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        config: req.app.config,
+        menu: sortMenu(await getMenu(db)),
+        csrfToken: req.csrfToken()
+    });
+});
+
+
+// pages editor
+const editPage=(async (req, res) => {
+    const db = req.app.db;
+    const page = await db.pages.findOne({ _id: getId(req.params.page) });
+    const menu = sortMenu(await getMenu(db));
+    if(!page){
+        res.status(404).render('error', {
+            title: '404 Error - Page not found',
+            config: req.app.config,
+            message: '404 Error - Page not found',
+            helpers: req.handlebars.helpers,
+            showFooter: 'showFooter',
+            menu
+        });
+        return;
+    }
+
+    res.render('settings-page', {
+        title: 'Static pages',
+        page: page,
+        button_text: 'Update',
+        session: req.session,
+        admin: true,
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        config: req.app.config,
+        menu,
+        csrfToken: req.csrfToken()
+    });
+});
+
+
+
+
+
+
+
+
+
+
+
+// insert/update page
+const insertPage=(async (req, res) => {
+    const db = req.app.db;
+
+    const doc = {
+        pageName: req.body.pageName,
+        pageSlug: req.body.pageSlug,
+        pageEnabled: req.body.pageEnabled,
+        pageContent: req.body.pageContent
+    };
+
+    if(req.body.pageId){
+        // existing page
+        const page = await db.pages.findOne({ _id: getId(req.body.pageId) });
+        if(!page){
+            res.status(400).json({ message: 'Page not found' });
+            return;
+        }
+
+        try{
+            const updatedPage = await db.pages.findOneAndUpdate({ _id: getId(req.body.pageId) }, { $set: doc }, { returnOriginal: false });
+            res.status(200).json({ message: 'Page updated successfully', pageId: req.body.pageId, page: updatedPage.value });
+        }catch(ex){
+            res.status(400).json({ message: 'Error updating page. Please try again.' });
+        }
+    }else{
+        // insert page
+        try{
+            const newDoc = await db.pages.insertOne(doc);
+            res.status(200).json({ message: 'New page successfully created', pageId: newDoc.insertedId });
+            return;
+        }catch(ex){
+            res.status(400).json({ message: 'Error creating page. Please try again.' });
+        }
+    }
+});
+
+// delete a page
+const deletePage=(async (req, res) => {
+    const db = req.app.db;
+
+    const page = await db.pages.findOne({ _id: getId(req.body.pageId) });
+    if(!page){
+        res.status(400).json({ message: 'Page not found' });
+        return;
+    }
+
+    try{
+        await db.pages.deleteOne({ _id: getId(req.body.pageId) }, {});
+        res.status(200).json({ message: 'Page successfully deleted' });
+        return;
+    }catch(ex){
+        res.status(400).json({ message: 'Error deleting page. Please try again.' });
+    }
+});
+
+// new menu item
+const menuItem=( (req, res) => {
+    const result = newMenu(req);
+    if(result === false){
+        res.status(400).json({ message: 'Failed creating menu.' });
+        return;
+    }
+    res.status(200).json({ message: 'Menu created successfully.' });
+});
+
+// update existing menu item
+const updatemenu=( (req, res) => {
+    const result = updateMenu(req);
+    if(result === false){
+        res.status(400).json({ message: 'Failed updating menu.' });
+        return;
+    }
+    res.status(200).json({ message: 'Menu updated successfully.' });
+});
+
+// delete menu item
+const deletemenu=((req, res) => {
+    const result = deleteMenu(req, req.body.menuId);
+    if(result === false){
+        res.status(400).json({ message: 'Failed deleting menu.' });
+        return;
+    }
+    res.status(200).json({ message: 'Menu deleted successfully.' });
+});
+
+// We call this via a Ajax call to save the order from the sortable list
+const saveOrder=( (req, res) => {
+    const result = orderMenu(req, res);
+    if(result === false){
+        res.status(400).json({ message: 'Failed saving menu order' });
+        return;
+    }
+    res.status(200).json({});
+});
+
+
+
+
+
+
+
+
+
+
+
+
+const validatePerma=(async (req, res) => {
+    // if doc id is provided it checks for permalink in any products other that one provided,
+    // else it just checks for any products with that permalink
+    const db = req.app.db;
+
+    let query = {};
+    if(typeof req.body.docId === 'undefined' || req.body.docId === ''){
+        query = { productPermalink: req.body.permalink };
+    }else{
+        query = { productPermalink: req.body.permalink, _id: { $ne: getId(req.body.docId) } };
+    }
+
+    const products = await db.products.countDocuments(query);
+    if(products && products > 0){
+        res.status(400).json({ message: 'Permalink already exists' });
+        return;
+    }
+    res.status(200).json({ message: 'Permalink validated successfully' });
+});
+
+// Discount codes
+const discount=(async (req, res) => {
+    const db = req.app.db;
+
+    const discounts = await db.discounts.find({}).toArray();
+
+    res.render('settings-discounts', {
+        title: 'Discount code',
+        config: req.app.config,
+        session: req.session,
+        discounts,
+        admin: true,
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        csrfToken: req.csrfToken()
+    });
+});
+
+// Edit a discount code
+const editDiscount=( async (req, res) => {
+    const db = req.app.db;
+
+    const discount = await db.discounts.findOne({ _id: getId(req.params.id) });
+
+    res.render('settings-discount-edit', {
+        title: 'Discount code edit',
+        session: req.session,
+        admin: true,
+        discount,
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        config: req.app.config,
+        csrfToken: req.csrfToken()
+    });
+});
+
+// Update discount code
+const updateDiscount=(async (req, res) => {
+    const db = req.app.db;
+
+     // Doc to insert
+     const discountDoc = {
+        discountId: req.body.discountId,
+        code: req.body.code,
+        type: req.body.type,
+        value: parseInt(req.body.value),
+        start: moment(req.body.start, 'DD/MM/YYYY HH:mm').toDate(),
+        end: moment(req.body.end, 'DD/MM/YYYY HH:mm').toDate()
+    };
+
+    // Validate the body again schema
+    const schemaValidate = validateJson('editDiscount', discountDoc);
+    if(!schemaValidate.result){
+        res.status(400).json(schemaValidate.errors);
+        return;
+    }
+
+    // Check start is after today
+    if(moment(discountDoc.start).isBefore(moment())){
+        res.status(400).json({ message: 'Discount start date needs to be after today' });
+        return;
+    }
+
+    // Check end is after the start
+    if(!moment(discountDoc.end).isAfter(moment(discountDoc.start))){
+        res.status(400).json({ message: 'Discount end date needs to be after start date' });
+        return;
+    }
+
+    // Check if code exists
+    const checkCode = await db.discounts.countDocuments({
+        code: discountDoc.code,
+        _id: { $ne: getId(discountDoc.discountId) }
+    });
+    if(checkCode){
+        res.status(400).json({ message: 'Discount code already exists' });
+        return;
+    }
+
+    // Remove discountID
+    delete discountDoc.discountId;
+
+    try{
+        await db.discounts.updateOne({ _id: getId(req.body.discountId) }, { $set: discountDoc }, {});
+        res.status(200).json({ message: 'Successfully saved', discount: discountDoc });
+    }catch(ex){
+        res.status(400).json({ message: 'Failed to save. Please try again' });
+    }
+});
+
+// Create a discount code
+const getDiscount=( async (req, res) => {
+    res.render('settings-discount-new', {
+        title: 'Discount code create',
+        session: req.session,
+        admin: true,
+        message: clearSessionValue(req.session, 'message'),
+        messageType: clearSessionValue(req.session, 'messageType'),
+        helpers: req.handlebars.helpers,
+        config: req.app.config,
+        csrfToken: req.csrfToken()
+    });
+});
+
+// Create a discount code
+const createDiscount=( async (req, res) => {
+    const db = req.app.db;
+
+    // Doc to insert
+    const discountDoc = {
+        code: req.body.code,
+        type: req.body.type,
+        value: parseInt(req.body.value),
+        start: moment(req.body.start, 'DD/MM/YYYY HH:mm').toDate(),
+        end: moment(req.body.end, 'DD/MM/YYYY HH:mm').toDate()
+    };
+
+    // Validate the body again schema
+    const schemaValidate = validateJson('newDiscount', discountDoc);
+    if(!schemaValidate.result){
+        res.status(400).json(schemaValidate.errors);
+        return;
+    }
+
+    // Check if code exists
+    const checkCode = await db.discounts.countDocuments({
+        code: discountDoc.code
+    });
+    if(checkCode){
+        res.status(400).json({ message: 'Discount code already exists' });
+        return;
+    }
+
+    // Check start is after today
+    if(moment(discountDoc.start).isBefore(moment())){
+        res.status(400).json({ message: 'Discount start date needs to be after today' });
+        return;
+    }
+
+    // Check end is after the start
+    if(!moment(discountDoc.end).isAfter(moment(discountDoc.start))){
+        res.status(400).json({ message: 'Discount end date needs to be after start date' });
+        return;
+    }
+
+    // Insert discount code
+    const discount = await db.discounts.insertOne(discountDoc);
+    res.status(200).json({ message: 'Discount code created successfully', discountId: discount.insertedId });
+});
+
+// Delete discount code
+const deleteDiscount=(async (req, res) => {
+    const db = req.app.db;
+
+    try{
+        await db.discounts.deleteOne({ _id: getId(req.body.discountId) }, {});
+        res.status(200).json({ message: 'Discount code successfully deleted' });
+        return;
+    }catch(ex){
+        res.status(400).json({ message: 'Error deleting discount code. Please try again.' });
+    }
+});
+
+// Add image by URL
+const addImage=(async (req, res) => {
+    const db = req.app.db;
+
+    // get the product form the DB
+    const product = await db.products.findOne({ _id: getId(req.body.productId) });
+    if(!product){
+        // Return error
+        res.status(400).json({ message: 'Image error. Please try again.' });
+        return;
+    }
+
+    // Check image URL already in list
+    if(product.productImages){
+        if(product.productImages.includes(req.body.imageUrl)){
+            res.status(400).json({ message: 'Image error. Image with that URL already exists.' });
+            return;
+        }
+    }
+
+    // Check image URL already set as main image
+    if(product.productImage === req.body.imageUrl){
+        res.status(400).json({ message: 'Image error. Image with that URL already exists.' });
+        return;
+    }
+
+    // Check productImages and init
+    if(!product.productImages){
+        product.productImages = [];
+    }
+    // Add the image to our images
+    product.productImages.push(req.body.imageUrl);
+
+    try{
+        // if there isn't a product featured image, set this one
+        if(!product.productImage){
+            await db.products.updateOne({ _id: getId(req.body.productId) }, { $set: { productImage: req.body.imageUrl } }, { multi: false });
+        }
+
+        // Add the images
+        await db.products.updateOne({ _id: getId(req.body.productId) }, { $set: { productImages: product.productImages } }, { multi: false });
+        res.status(200).json({ message: 'Image added successfully' });
+    }catch(ex){
+        console.log('Failed to upload the file', ex);
+        res.status(400).json({ message: 'Image error. Please try again.' });
+    }
+});
+
+// upload the file
+// const upload = multer({ dest: 'public/uploads/' });
+const uploadFile=(async (req, res) => {
+    const db = req.app.db;
+
+    if(req.file){
+        const file = req.file;
+
+        // Get the mime type of the file
+        const mimeType = mime.lookup(file.originalname);
+
+        // Check for allowed mime type and file size
+        if(!allowedMimeType.includes(mimeType) || file.size > fileSizeLimit){
+            // Remove temp file
+            fs.unlinkSync(file.path);
+
+            // Return error
+            res.status(400).json({ message: 'File type not allowed or too large. Please try again.' });
+            return;
+        }
+
+        // get the product form the DB
+        const product = await db.products.findOne({ _id: getId(req.body.productId) });
+        if(!product){
+            // delete the temp file.
+            fs.unlinkSync(file.path);
+
+            // Return error
+            res.status(400).json({ message: 'File upload error. Please try again.' });
+            return;
+        }
+
+        const productPath = product._id.toString();
+        const uploadDir = path.join('public/uploads', productPath);
+
+        // Check directory and create (if needed)
+        checkDirectorySync(uploadDir);
+
+        // Setup the new path
+        const imagePath = path.join('/uploads', productPath, file.originalname.replace(/ /g, '_'));
+
+        // save the new file
+        const dest = fs.createWriteStream(path.join(uploadDir, file.originalname.replace(/ /g, '_')));
+        const pipeline = util.promisify(stream.pipeline);
+
+        try{
+            await pipeline(
+                fs.createReadStream(file.path),
+                dest
+            );
+
+            // delete the temp file.
+            fs.unlinkSync(file.path);
+
+            // if there isn't a product featured image, set this one
+            if(!product.productImage){
+                await db.products.updateOne({ _id: getId(req.body.productId) }, { $set: { productImage: imagePath } }, { multi: false });
+            }
+            res.status(200).json({ message: 'File uploaded successfully' });
+        }catch(ex){
+            console.log('Failed to upload the file', ex);
+            res.status(400).json({ message: 'File upload error. Please try again.' });
+        }
+    }else{
+        // Return error
+        console.log('fail', req.file);
+        res.status(400).json({ message: 'File upload error. Please try again.' });
+    }
+});
+
+// delete a file via ajax request
+const testEmail=((req, res) => {
+    const config = req.app.config;
+    // TODO: Should fix this to properly handle result
+    sendEmail(config.emailAddress, 'expressCart test email', 'Your email settings are working');
+    res.status(200).json({ message: 'Test email sent' });
+});
+
+const searchAll=(async (req, res, next) => {
+    const db = req.app.db;
+    const searchValue = req.body.searchValue;
+    const limitReturned = 5;
+
+    // Empty arrays
+    let customers = [];
+    let orders = [];
+    let products = [];
+
+    // Default queries
+    const customerQuery = {};
+    const orderQuery = {};
+    const productQuery = {};
+
+    // If an ObjectId is detected use that
+    if(ObjectId.isValid(req.body.searchValue)){
+        // Get customers
+        customers = await db.customers.find({
+            _id: ObjectId(searchValue)
+        })
+        .limit(limitReturned)
+        .sort({ created: 1 })
+        .toArray();
+
+        // Get orders
+        orders = await db.orders.find({
+            _id: ObjectId(searchValue)
+        })
+        .limit(limitReturned)
+        .sort({ orderDate: 1 })
+        .toArray();
+
+        // Get products
+        products = await db.products.find({
+            _id: ObjectId(searchValue)
+        })
+        .limit(limitReturned)
+        .sort({ productAddedDate: 1 })
+        .toArray();
+
+        return res.status(200).json({
+            customers,
+            orders,
+            products
+        });
+    }
+
+    // If email address is detected
+    if(emailRegex.test(req.body.searchValue)){
+        customerQuery.email = searchValue;
+        orderQuery.orderEmail = searchValue;
+    }else if(numericRegex.test(req.body.searchValue)){
+        // If a numeric value is detected
+        orderQuery.amount = req.body.searchValue;
+        productQuery.productPrice = req.body.searchValue;
+    }else{
+        // String searches
+        customerQuery.$or = [
+            { firstName: { $regex: new RegExp(searchValue, 'img') } },
+            { lastName: { $regex: new RegExp(searchValue, 'img') } }
+        ];
+        orderQuery.$or = [
+            { orderFirstname: { $regex: new RegExp(searchValue, 'img') } },
+            { orderLastname: { $regex: new RegExp(searchValue, 'img') } }
+        ];
+        productQuery.$or = [
+            { productTitle: { $regex: new RegExp(searchValue, 'img') } },
+            { productDescription: { $regex: new RegExp(searchValue, 'img') } }
+        ];
+    }
+
+    // Get customers
+    if(Object.keys(customerQuery).length > 0){
+        customers = await db.customers.find(customerQuery)
+        .limit(limitReturned)
+        .sort({ created: 1 })
+        .toArray();
+    }
+
+    // Get orders
+    if(Object.keys(orderQuery).length > 0){
+        orders = await db.orders.find(orderQuery)
+        .limit(limitReturned)
+        .sort({ orderDate: 1 })
+        .toArray();
+    }
+
+    // Get products
+    if(Object.keys(productQuery).length > 0){
+        products = await db.products.find(productQuery)
+        .limit(limitReturned)
+        .sort({ productAddedDate: 1 })
+        .toArray();
+    }
+
+    return res.status(200).json({
+        customers,
+        orders,
+        products
+    });
+});
+
+
+
+
 module.exports={adminDashboard,
     settingsMenu,
      logout, login, 
@@ -270,5 +885,27 @@ module.exports={adminDashboard,
       dashboard, 
       getSettings,
       genAPI, 
-      settingsUpdate
+      settingsUpdate,
+      getPages,
+      newPages,
+      editPage,
+      insertPage,
+      deletePage,
+      menuItem,
+      updatemenu,
+      deletemenu,
+      saveOrder,
+      testEmail,
+      searchAll,
+      validatePerma,
+    discount,
+    editDiscount,
+    updateDiscount,
+    getDiscount,
+    createDiscount,
+    deleteDiscount,
+    addImage,
+    uploadFile,
+
+
     }
